[update] ブザーの引数を変更

+++ Core/App/Module/UI.hpp
@@ -15,7 +15,7 @@ class UI {
     public:
         UI(MAL* mcu);
         void init();
-        void buzzer(uint16_t pulse, uint8_t tim);
+        void buzzer(float pulse, uint8_t tim);
         uint8_t getRotarySW();
         bool getSW();
         void setLED(uint8_t pin, bool states);
